Extract ProtoHelpersFuncMap in its own file
